backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6
requests>=2.31.0
python-dotenv>=1.0.0
replicate>=0.25.0
genesis-world==0.3.7
openai>=1.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=5.0.0
structlog>=23.2.0
tenacity>=8.2.3
pydantic-settings>=2.1.0
anthropic>=0.7.7
pillow>=10.1.0
numpy>=1.26.4
opencv-python>=4.8.1.78
slowapi>=0.1.9
redis>=5.0.0
